backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
openai>=1.0.0
pillow>=10.0.0
PyPDF2>=3.0.0
pandas>=2.0.0
networkx>=3.0
reportlab>=4.0.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.9
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openpyxl>=3.1.0
xlrd>=2.0.1
chardet>=5.0.0
python-dateutil>=2.8.2

# Production server
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0

# Optional: python-magic (requires libmagic on Windows)
# For better MIME type detection, install separately if needed
# python-magic>=0.4.27
# python-magic-bin>=0.4.14  # Windows only

